requirements.txt
# utils for resume parsing
python-docx
nltk
spacy==2.3.5
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
pyresparser
pandas
requests

# GUI
PyQt5
customtkinter
pillow


# for type hinting using pydantic and dataclasses
pydantic
dataclasses


# SQL database
sqlalchemy


python-dotenv